Completed CRUD and objects  chage for Challenge
